chore: re-organize mod `lmkcdey`
