+++ crypto/src/scheme/blind_rotation/lmkcdey.rs
@@ -1,13 +1,10 @@
 mod method;
 mod structure;
 
-#[cfg(test)]
-mod test;
+pub mod interactive;
 
-pub use method::{
-    aggregate_bs_key_shares, aggregate_pk_shares, bootstrap, bs_key_gen, bs_key_share_gen,
-    prepare_bs_key,
-};
-pub use structure::{
-    LmkcdeyInteractiveCrs, LmkcdeyInteractiveParam, LmkcdeyKey, LmkcdeyKeyShare, LmkcdeyParam,
-};
+#[cfg(test)]
+pub(crate) mod test;
+
+pub use method::{bootstrap, bs_key_gen, prepare_bs_key};
+pub use structure::{LmkcdeyKey, LmkcdeyParam};
